+++ src/main/java/appeng/core/localization/GuiText.java
@@ -127,6 +127,7 @@ public enum GuiText {
     EncodedBy,
     Substitute,
     UsesFluidsDirectly,
+    ViewPatternHint,
     Yes,
     No,